rename NavBar to Navbar
