+++ src/components/rootLayout/RootLayout.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { Outlet } from "react-router-dom";
-import NavBar from "../NavBar";
+import NavBar from "../Navbar";
 import Footer from "../Footer";
 import ScrollToTop from "../ui/ScrollToTop ";
